Bump packages, minor UI cleanup
