+++ package.json
@@ -22,7 +22,7 @@
     "@mantine/notifications": "^7.16.3",
     "@maverick-js/signals": "^6.0.0",
     "@tabler/icons-react": "^3.30.0",
-    "@tanstack/react-router": "^1.99.13",
+    "@tanstack/react-router": "^1.102.5",
     "@types/downloadjs": "^1.4.6",
     "clsx": "2.1.1",
     "downloadjs": "^1.4.7",
@@ -40,19 +40,19 @@
   },
   "devDependencies": {
     "@biomejs/biome": "1.9.4",
-    "@tanstack/router-cli": "^1.99.14",
-    "@tanstack/router-plugin": "^1.99.14",
-    "@types/node": "^22.13.1",
+    "@tanstack/router-cli": "^1.102.6",
+    "@tanstack/router-plugin": "^1.102.6",
+    "@types/node": "^22.13.2",
     "@types/react": "^19.0.8",
     "@types/react-dom": "^19.0.3",
     "@vitejs/plugin-legacy": "^6.0.1",
     "@vitejs/plugin-react": "^4.3.4",
     "autoprefixer": "^10.4.20",
-    "globals": "^15.14.0",
-    "postcss": "^8.5.1",
+    "globals": "^15.15.0",
+    "postcss": "^8.5.2",
     "postcss-preset-mantine": "^1.17.0",
     "postcss-simple-vars": "^7.0.1",
-    "terser": "^5.38.1",
+    "terser": "^5.39.0",
     "typescript": "^5.7.3",
     "vite": "^6.1.0",
     "vite-plugin-pwa": "^0.21.1"
